+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Version 3.5.0-DEV
 -------------
 - ![FEATURE][badge-feature] `sasa` function to compute solvent accessible surface area.
 - ![FEATURE][badge-feature] Add `vdw_radius` as element property, and `element_vdw_radius` function to fetch them.
+- ![INFO][badge-info] The element symbol is returned as a `InlineStrings.String3`. 
 
 Version 3.4.0
 -------------